+++ damus/Models/SearchHomeModel.swift
@@ -30,7 +30,7 @@ class SearchHomeModel: ObservableObject {
     func get_base_filter() -> NostrFilter {
         var filter = NostrFilter(kinds: [.text, .chat])
         filter.limit = self.limit
-        filter.until = Int64(Date.now.timeIntervalSince1970)
+        filter.until = UInt32(Date.now.timeIntervalSince1970)
         return filter
     }
